new SchemaDef plugin type to define site-specific schemas (#342)
* new SchemaDef plugin type to define site-specific schemas
* add missing _update_plugin_source for pkg_resources; force schemadefs to load
* add SchemaDef plugin test
* add method to extend all three manifest plugin lists at once
* cleaner save/restore of existing _MANIFEST
* schemadef plugin classes added to otio.schemadef namespace
* Explicitly include our source files in MANIFEST.in

Travis-CI is having some problems reconciling the file list between
sdist and the repo and suggested adding this line:
recursive-include opentimelineio *
so I'm adding it to see if that works.
The particular problem is that it is not picking up one of my latest
additions, namely the schemadef directory and the __init__.py file
in it.  That's the only source file in that directory -- maybe that's
the problem?

* add UnknownSchema object to pass through undefined schemas

Old behavior was to through an exception if any object being read
(or created by instance_from_schema) had an unregistered schema name.
New behavior is to replace the schema of such objects with the
UnknownSchema schema type and save the original schema label in a
field called "UnknownSchemaOriginalLabel".  Such objects are then
available within the OTIO data structure.  The new property
"is_unknown_schema" will be True for these objects and False for all
other SerializableObjects.  When an UnknownSchema is serialized,
the json_serializer will replace the UnknownSchema label with the
original label, so that (for instance) otiocat will pass such
unregistered objects through unchanged.  This is part of the schemadef
project because schemadef plugins imply that we will start to see
site or studio-specific OTIO schema types, and it would be more polite
to pass these through quietly if you don't understand them.

* improve docs and tests for schemadef and unknown schema