Optimizations for flatten (range_of_all_children) (#358)
* Add range-map precompute function.

schema.Track.range_of_child is expensive for tracks with many items
because it searches through each previous child to compute the duration
prior to item passed in.  When iterating over a track, this function
lets you precompute that map to get the ranges in one shot much more
effeciently than the old way.

For operations like flattening, this is a huge performance win.